Fix user mention username formatting for deleted users
When user is mentioned in a page and later that user is deleted, the
`data-account-id` exists but that user doesn't exist in the Confluence instance
anymore and `User.from_accountid` will raise an exception.

Changed the implementation to fall back to the username extraction from the
string if the user is not found.